cli/pyproject.toml
[project]
name = "downme-cli"
version = "0.0.1"
dependencies = ["click", "requests", "azure-storage-blob==12.9.1", "azure-core==1.29.0"]

[project.scripts]
downme = "downme:main"